instock_connector/requirements.txt
inorbit-edge[video]>=1.13.0,<2.0
pydantic>=2.6,<3.0
pytz>=2024.1
pyyaml>=6.0,<7.0
requests>=2.31,<3.0
